PagePilot/requirements.txt
tiktoken
faiss-cpu
<<<<<<< HEAD
langchain==0.0.200
=======
langchain
>>>>>>> 9d3f538c47585bb8cc75c008e04857a14cbd826a
PyPDF2
python-dotenv
streamlit
Pillow
sentence-transformers
google-generativeai